Move SceneAwareComboBox to ui/controls subpackage
Extract the combobox subclass out of param_widgets.py into its own
module under a new src/ui/controls/ folder so reusable custom controls
have a home separate from node param editor widgets.
