+++ src/ui/param_widgets.py
@@ -9,9 +9,7 @@ from typing_extensions import override
 from PySide6.QtCore import Qt, Signal
 from PySide6.QtWidgets import (
     QCheckBox,
-    QComboBox,
     QFileDialog,
-    QGraphicsItem,
     QHBoxLayout,
     QLineEdit,
     QPushButton,
@@ -21,6 +19,7 @@ from PySide6.QtWidgets import (
 
 from constants import INPUT_DIR, OUTPUT_DIR
 from core.node_base import NodeBase, NodeParam, NodeParamType
+from ui.controls.scene_aware_combobox import SceneAwareComboBox
 
 logger = logging.getLogger(__name__)
 
@@ -141,62 +140,6 @@ class BoolParamWidget(ParamWidgetBase):
         return self._check.isChecked()
 
 
-class _SceneAwareComboBox(QComboBox):
-    """QComboBox that keeps its popup above overlapping graphics items.
-
-    When a QComboBox is embedded via QGraphicsProxyWidget, Qt renders its
-    popup as a child proxy of the host proxy widget. Two Z-order issues
-    arise while the popup is visible:
-
-    * Sibling NodeItems at the same Z can occlude the popup (inter-node).
-    * The popup extends below its host proxy, and sibling child items of
-      the NodeItem with higher Z (e.g. the close button, ports) paint on
-      top of it (intra-node).
-
-    While the popup is open we boost the host NodeItem's Z above other
-    nodes AND the params proxy's Z above its NodeItem siblings, then
-    restore both on close.
-    """
-
-    _POPUP_Z_BOOST: float = 10_000.0
-
-    def __init__(self, parent: QWidget | None = None) -> None:
-        super().__init__(parent)
-        self._raised_proxy: QGraphicsItem | None = None
-        self._raised_node: QGraphicsItem | None = None
-        self._saved_proxy_z: float | None = None
-        self._saved_node_z: float | None = None
-
-    @override
-    def showPopup(self) -> None:
-        if self._raised_node is None:
-            proxy = self.window().graphicsProxyWidget()
-            if proxy is not None:
-                node = proxy
-                while node.parentItem() is not None:
-                    node = node.parentItem()
-                self._raised_node = node
-                self._saved_node_z = node.zValue()
-                node.setZValue(self._saved_node_z + self._POPUP_Z_BOOST)
-                if proxy is not node:
-                    self._raised_proxy = proxy
-                    self._saved_proxy_z = proxy.zValue()
-                    proxy.setZValue(self._saved_proxy_z + self._POPUP_Z_BOOST)
-        super().showPopup()
-
-    @override
-    def hidePopup(self) -> None:
-        super().hidePopup()
-        if self._raised_proxy is not None and self._saved_proxy_z is not None:
-            self._raised_proxy.setZValue(self._saved_proxy_z)
-        if self._raised_node is not None and self._saved_node_z is not None:
-            self._raised_node.setZValue(self._saved_node_z)
-        self._raised_proxy = None
-        self._raised_node = None
-        self._saved_proxy_z = None
-        self._saved_node_z = None
-
-
 class EnumParamWidget(ParamWidgetBase):
     """Combo-box editor for :attr:`NodeParamType.ENUM` parameters.
 
@@ -220,7 +163,7 @@ class EnumParamWidget(ParamWidgetBase):
             )
         self._enum_cls: type[Enum] = enum_cls
 
-        self._combo = _SceneAwareComboBox()
+        self._combo = SceneAwareComboBox()
         self._combo.setMinimumWidth(96)
         for member in self._enum_cls:
             self._combo.addItem(self._label_for(member), member)
